requirements.txt
fastapi
fastapi[standard]
openai
requests
prompty[azure]==0.1.24
promptflow
promptflow-core
promptflow-evals
pandas
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-fastapi
jsonlines
python-dotenv==1.0.1
azure-search-documents==11.4.0
gunicorn==21.2.0
jupyter
opentelemetry-instrumentation
azure-identity==1.17.1
youtube-transcript-api
azure-cosmos
PyJWT
keyrings.alt